BSP: Typecast isdigit argument with unsigned char
Typecast isdigit argument with unsigned char instead of
signed as isdigit function expects input as unsigned
character and fix below misrac violation.

misra_c_2012_rule_21_13_violation: The value _ctype_ + 1U +
(int)(s32)*cptr of type char const *, which is passed to a
function in <ctype.h>, shall be representable as an unsigned
char or be the value EOF.


Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.paladugu@xilinx.com>

+++ lib/bsp/standalone/src/common/xil_printf.c
@@ -209,7 +209,7 @@ static s32 getnum(charptr* linep)
 	charptr cptr = *linep;
 
 	while (cptr != NULL) {
-		ResultIsDigit = isdigit(((s32)*cptr));
+		ResultIsDigit = isdigit(((u8)*cptr));
 		if (ResultIsDigit == 0) {
 			break;
 		}
